Added button to manifest powers
There is now a button on the right side that subtracts the cost of a selected power (and any bonus augmentation points spent) from the character's current power points. There is a slider that lets the user choose how hard to augment their powers. It does limit you from augmenting a power's cost above the character's manifester level.